+++ Assets/Bean/PlayerInfo.cs
@@ -1,4 +1,6 @@
+using System;
 using Photon.Bolt;
+using UnityEngine;
 
 public class PlayerInfo : IProtocolToken
 {
@@ -12,5 +14,19 @@ public class PlayerInfo : IProtocolToken
     public void Read(UdpKit.UdpPacket packet)
     {
         name = packet.ReadString();
+    }
+
+    public static int MaxNameLength = 16;
+
+    public static string SavedName
+    {
+        get
+        {
+            return PlayerPrefs.GetString("name", "Player Name").Truncate(MaxNameLength);
+        }
+        set
+        {
+            PlayerPrefs.SetString("name", value.Truncate(MaxNameLength));
+        }
     }
 }